# feat: use native `realpath` if available to unwrap symlinks