--- a/lib/async.js
+++ b/lib/async.js
@@ -5,6 +5,8 @@ var nodeModulesPaths = require('./node-modules-paths.js');
 var normalizeOptions = require('./normalize-options.js');
 var isCore = require('./is-core');
 
+var realpath = typeof fs.realpath.native === 'function' ? fs.realpath.native : fs.realpath;
+
 var defaultIsFile = function isFile(file, cb) {
     fs.stat(file, function (err, stat) {
         if (!err) {
@@ -27,7 +29,7 @@ var defaultIsDir = function isDirectory(dir, cb) {
 
 var maybeUnwrapSymlink = function maybeUnwrapSymlink(x, opts, cb) {
     if (!opts || !opts.preserveSymlinks) {
-        fs.realpath(x, function (realPathErr, realPath) {
+        realpath(x, function (realPathErr, realPath) {
             if (realPathErr && realPathErr.code !== 'ENOENT') cb(realPathErr);
             else cb(null, realPathErr ? x : realPath);
         });