--- a/lib/sync.js
+++ b/lib/sync.js
@@ -5,6 +5,8 @@ var caller = require('./caller.js');
 var nodeModulesPaths = require('./node-modules-paths.js');
 var normalizeOptions = require('./normalize-options.js');
 
+var realpath = typeof fs.realpathSync.native === 'function' ? fs.realpathSync.native : fs.realpathSync;
+
 var defaultIsFile = function isFile(file) {
     try {
         var stat = fs.statSync(file);
@@ -28,7 +30,7 @@ var defaultIsDir = function isDirectory(dir) {
 var maybeUnwrapSymlink = function maybeUnwrapSymlink(x, opts) {
     if (!opts || !opts.preserveSymlinks) {
         try {
-            return fs.realpathSync(x);
+            return realpath(x);
         } catch (realPathErr) {
             if (realPathErr.code !== 'ENOENT') {
                 throw realPathErr;