As discussed in https://github.com/facebook/jest/issues/9776

Docs: https://nodejs.org/api/fs.html#fs_fs_realpath_native_path_options_callback

didn't bother with `process.bindings` as we really shouldn't use those