@SimenB 2020-04-07
As an aside, it might make sense to make `maybeUnwrapSymlink` pluggable like `readFile`, `isFile` and `isDirectory` is already. Those 3 allow you to avoid the real fs if necessary, but the `realpath` call does not. I can open up a separate issue for that?