@ljharb 2020-04-07
A separate issue (or even a PR) sounds good.