@SimenB 2020-04-07
On my machine `npmt test` seems a _tiny_ bit faster, not sure how much of a benchmark that is, though. probably not at all 😀 

Master:
```
  Time (mean ± σ):      6.358 s ±  0.548 s    [User: 7.007 s, System: 0.792 s]
  Range (min … max):    5.776 s …  7.328 s    10 runs
```

This branch
```
  Time (mean ± σ):      6.243 s ±  0.390 s    [User: 6.979 s, System: 0.751 s]
  Range (min … max):    5.929 s …  6.992 s    10 runs
```