@ljharb 2020-04-07
when all the optionals are done, we can compare the "total time" on the travis builds :-p