@ljharb 2020-04-08
Looks like travis took 3:34:08 (214.13 min) on master, and 2:33:30 (153.5 min) on the PR, across 176 jobs, so that's about 28% faster, or about 1min 13s vs 52s - seems like an improvement to me!

Appveyor needed a few reruns, so we'll see how that shakes out.