@SimenB 2020-04-08
@ljharb one thing - using the native `realpath` might cause issues, see https://github.com/nodejs/node/issues/7175

From what I gather `require` uses the native `realpath` (although I haven't verified by looking at source code) so I still think it's safe (or even correct) to change, but a heads up seems warranted in any case