@ljharb 2020-04-08
It looks like perhaps using https://www.npmjs.com/package/fs.realpath would handle those issues, when using the native realpath?